+++ app/src/assets/localization/en/labware_position_check.json
@@ -18,9 +18,9 @@
   "labware_step_detail_modal_nozzle_or_tip": "To ensure the nozzle or tip is level with the top of the labware, position yourself at eye-level and/or slide a sheet of paper between the nozzle and tip.",
   "labware_step_detail_modal_nozzle_image_1_text": "Viewed from front, it appears centered...",
   "labware_step_detail_modal_nozzle_image_2_text": "...but viewed from side, it requires adjustment",
-  "labware_step_detail_modal_nozzle_image_2_nozzletext": "Nozzle is not centered",
+  "labware_step_detail_modal_nozzle_image_2_nozzle_text": "Nozzle is not centered",
   "labware_step_detail_modal_nozzle_or_tip_image_1_text": "Viewed from standing height, it appears level...",
   "labware_step_detail_modal_nozzle_or_tip_image_2_text": "... but viewed from eye-level, it requires adjustment",
-  "labware_step_detail_modal_nozzle_or_tip_image_2_nozzletext": "Nozzle is not level",
+  "labware_step_detail_modal_nozzle_or_tip_image_2_nozzle_text": "Nozzle is not level",
   "labware_step_detail_modal_nozzle_or_tip_image_3_text": "If you’re having trouble, slide 1 sheet of printer paper between the nozzle and the tip. A single piece of paper should barely pass between them."
 }
