+++ app/src/organisms/ProtocolSetup/LabwarePositionCheck/__tests__/LabwarePositionCheckStepDetailModal.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { fireEvent } from '@testing-library/react'
-import { renderWithProviders } from '@opentrons/components/__utils__'
+import { renderWithProviders } from '@opentrons/components/src/testing/utils'
 import { i18n } from '../../../../i18n'
 import { LabwarePositionCheckStepDetailModal } from '../LabwarePositionCheckStepDetailModal'
 
@@ -12,7 +12,7 @@ const render = (
     {
       i18nInstance: i18n,
     }
-  )
+  )[0]
 }
 
 describe('LabwarePositionCheckStepDetailModal', () => {
